requirements.txt
aioredis==1.3.1
asgi-redis==1.4.3
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.2.1
Automat==20.2.0
autopep8==1.5
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
Django==3.0.2
django-appconf==1.0.4
django-autoslug-iplweb==1.9.5
django-avatar==5.0.0
django-bootstrap-modal-forms==1.5.0
django-dash==0.5.6
django-nine==0.2.3
django-tinymce==2.8.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
easy-thumbnails==2.7
feedparser==5.2.1
future==0.18.2
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
line-bot-sdk==1.16.0
msgpack==0.6.2
msgpack-python==0.5.6
ordereddict==1.1
pif==0.8.2
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
PyHamcrest==2.0.0
pyOpenSSL==19.1.0
pytz==2019.3
redis==2.10.6
requests==2.22.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
transliterate==1.10.2
Twisted==19.10.0
txaio==20.1.1
Unidecode==1.1.1
Unipath==1.1
urllib3==1.25.8
vishap==0.1.5
zope.interface==4.7.1
